@@ -1,7 +1,7 @@
 package com.someoctets.palets;
 
 import android.content.Context;
-import android.support.design.widget.TextInputEditText;
+import com.google.android.material.textfield.TextInputEditText;
 import android.util.AttributeSet;
 
 import java.util.ArrayList;
